refactor(duplication): reorganise files
- split multiple components in one file into multiple files
