--- app/views/course/object_duplications/data.json.jbuilder
@@ -1,2 +1,0 @@
-# frozen_string_literal: true
-json.partial! 'course_duplication_data'
