+++ client/app/bundles/course/duplication/pages/Duplication/DuplicateItemsConfirmation/Surveys/index.tsx
@@ -1,45 +1,13 @@
-import {
-  Card,
-  CardContent,
-  Checkbox,
-  FormControlLabel,
-  ListSubheader,
-  Typography,
-} from '@mui/material';
-import TypeBadge from 'course/duplication/components/TypeBadge';
-import UnpublishedIcon from 'course/duplication/components/UnpublishedIcon';
+import { FC } from 'react';
+import { Card, CardContent, ListSubheader } from '@mui/material';
+
 import { duplicableItemTypes } from 'course/duplication/constants';
 import { selectDuplicationStore } from 'course/duplication/selectors/destinationInstance';
-import { Survey } from 'course/duplication/types';
 import { defaultComponentTitles } from 'course/translations.intl';
 import { useAppSelector } from 'lib/hooks/store';
 import useTranslation from 'lib/hooks/useTranslation';
-import { FC } from 'react';
 
-const { SURVEY } = duplicableItemTypes;
-
-interface RowProps {
-  survey: Survey;
-}
-
-const RowComponent: FC<RowProps> = (props) => {
-  const { survey } = props;
-
-  return (
-    <FormControlLabel
-      key={`survey_${survey.id}`}
-      control={<Checkbox checked />}
-      label={
-        <span className="flex flex-row items-center">
-          <TypeBadge itemType={duplicableItemTypes.SURVEY} />
-          <UnpublishedIcon tooltipId="itemUnpublished" />
-          <Typography className="font-bold">{survey.title}</Typography>
-        </span>
-      }
-      className="flex items-center w-auto"
-    />
-  );
-};
+import SurveyRow from './SurveyRow';
 
 const SurveyListing: FC = () => {
   const { t } = useTranslation();
@@ -65,7 +33,7 @@ const SurveyListing: FC = () => {
       <Card>
         <CardContent>
           {selectedSurveys.map((survey) => (
-            <RowComponent survey={survey} />
+            <SurveyRow key={survey.id} survey={survey} />
           ))}
         </CardContent>
       </Card>
